lab9/requirements.txt
appdirs==1.4.4
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
packaging==20.3
Pillow==7.1.2
pooch==1.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.1
scipy==1.4.1
six==1.14.0
tifffile==2020.5.7
urllib3==1.25.9
